requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.4
pydantic==2.10.3
openai==1.57.4
